+++ airless/operator/google/storage.py
@@ -363,7 +363,15 @@ class BatchWriteProcessOperator(BaseEventOperator):
         self.gcs_hook.upload(local_filepath, get_config('GCS_BUCKET_LANDING_ZONE_LOADER'), f'{directory}/append')
         os.remove(local_filepath)
 
-        self.move_files(from_bucket, get_config('GCS_BUCKET_LANDING_ZONE_PROCESSED'), directory, files)
+        file_paths = [directory + '/' + f for f in files]
+
+        self.gcs_hook.move_files(
+            from_bucket=from_bucket,
+            files=file_paths,
+            to_bucket=get_config('GCS_BUCKET_LANDING_ZONE_PROCESSED'),
+            to_directory=directory,
+            rewrite=False
+        )
 
     def read_files(self, bucket, directory, files):
         file_contents = []
@@ -383,14 +391,6 @@ class BatchWriteProcessOperator(BaseEventOperator):
         local_filepath = self.file_hook.get_tmp_filepath('merged.ndjson', add_timestamp=True)
         self.file_hook.write(local_filepath=local_filepath, data=file_contents, use_ndjson=True)
         return local_filepath
-
-    def move_files(self, from_bucket, to_bucket, directory, files):
-        for f in files:
-            self.gcs_hook.move(
-                from_bucket=from_bucket,
-                from_prefix=f'{directory}/{f}',
-                to_bucket=to_bucket,
-                to_directory=directory)
 
 
 class BatchAggregateParquetFilesOperator(BaseEventOperator):