+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 
 **unreleased**
+- [Refactor] Move files in `GcsHook` using batch operations to copy and delete files in order to improve performance
 
 **v0.0.60**
 - [Bugfix] In `BatchWriteProcessParquetOperator` change `send_to_processed_move` to use redirect function